Lengthened log messahe variables and fixed precision problem
git-svn-id: https://svn.berlios.de/svnroot/repos/glimmer-cism/glimmer-cism/glimmer/trunk@408 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

+++ src/fortran/glint_example.f90
@@ -240,8 +240,8 @@ contains
     use netcdf
 
     character(*) :: filename,varname
-    real(dp),dimension(:,:,:),pointer :: array
-    real(dp),dimension(:),pointer,optional :: lons,lats
+    real(rk),dimension(:,:,:),pointer :: array
+    real(rk),dimension(:),pointer,optional :: lons,lats
     real,dimension(:),pointer,optional :: lonbound,latbound
     integer :: ncerr,ncid,i,varid
     real(rk) :: offset=0.0,scale=1.0
